// packages/lego/src/components/DatePicker/src/utils/types.ts
export type NDatePickerBaseSupportType =
  | 'year'
  | 'month'
  | 'week'
  | 'date'
  | 'datetime'
  | 'date-minutes'
  | 'date-seconds'
  | 'date-range'
  | 'month-range'
  | 'year-range'
  | 'datetime-range'
  | 'date-minutes-range'
  | 'date-seconds-range';

export type NDatePickerTypeFormat<T extends string> =
  | T
  | KebabCaseToCamelCase<T>
  | KebabCaseToLowerCase<T>;

/**
 * date-picker support the NDatePickerBaseSupportType, and its lowercase and camelcase.
 */
export type NDatePickerSupportType = NDatePickerTypeFormat<NDatePickerBaseSupportType>;

/**
 * Accept date(time) type
 * support basic type: string | number | Dayjs | Date | null | undefined
 * support dayjs plugin type:
 * - objectSupport(@import('dayjs/plugin/objectSupport'))
 * - arraySupport(@import('dayjs/plugin/arraySupport'))
 */
export type NDatePickerAccessDatetimeType = ConfigType;